requirements.txt
certifi==2020.6.20
neat-python==0.92
pygame==1.9.6
wincertstore==0.2
